% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_analysis.R
\name{calculate_w_statistics}
\alias{calculate_w_statistics}
\title{Calculate W Statistics}
\usage{
calculate_w_statistics(
  original_pvals,
  knockoff_pvals,
  original_coefs,
  knockoff_coefs,
  method = "signed_max"
)
}
\arguments{
\item{original_pvals}{P-values from original variables}

\item{knockoff_pvals}{List of p-values from knockoff variables}

\item{original_coefs}{Coefficients from original variables}

\item{knockoff_coefs}{List of coefficients from knockoff variables}

\item{method}{Method for calculating W statistics ("signed_max", "difference")}
}
\value{
Vector of W statistics
}
\description{
Calculate the W statistics for knockoff variable selection.
}
