% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_analysis.R
\name{cox_knockoff_screen}
\alias{cox_knockoff_screen}
\title{Cox Knockoff Screen}
\usage{
cox_knockoff_screen(
  genotypes,
  knockoffs,
  time,
  status,
  covariates,
  null_model = NULL,
  adjust_method = "bonferroni",
  alpha = 0.05
)
}
\arguments{
\item{genotypes}{Original genotype matrix (samples x SNPs)}

\item{knockoffs}{List of knockoff genotype matrices}

\item{time}{Vector of survival times}

\item{status}{Vector of event indicators}

\item{covariates}{Matrix or data frame of covariates}

\item{null_model}{Pre-fitted null model (optional)}

\item{adjust_method}{Method for multiple testing adjustment ("bonferroni", "BH", "BY")}

\item{alpha}{Significance level (default: 0.05)}
}
\value{
A list containing:
  \item{original_pvals}{P-values for original variables}
  \item{knockoff_pvals}{List of p-values for knockoff variables}
  \item{original_coefs}{Coefficients for original variables}
  \item{knockoff_coefs}{List of coefficients for knockoff variables}
  \item{w_stats}{W statistics for variable selection}
}
\description{
Perform association analysis using Cox regression with original and knockoff variables.
}
\examples{
\dontrun{
# Perform knockoff screening
results <- cox_knockoff_screen(
  genotypes = plink_data$genotypes,
  knockoffs = knockoffs,
  time = pheno_data$time,
  status = pheno_data$status,
  covariates = pheno_data[, c("age", "sex", "bmi")]
)
}

}
