% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_covariates}
\alias{example_covariates}
\title{Example Covariate Data}
\format{
A data frame with 20 rows and 5 variables:
\describe{
  \item{FID}{Family ID}
  \item{IID}{Individual ID}
  \item{age}{Age in years}
  \item{sex}{Sex (0 = male, 1 = female)}
  \item{bmi}{Body mass index}
}
}
\source{
Simulated data for package examples
}
\usage{
example_covariates
}
\description{
Covariate data for the example samples.
}
\keyword{datasets}
