% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genotypes}
\alias{example_genotypes}
\title{Example Genotype Data}
\format{
A 20 x 50 sparse matrix (dgCMatrix)
}
\source{
Simulated data for package examples
}
\usage{
example_genotypes
}
\description{
A sparse matrix containing example genotype data for 20 samples and 50 SNPs.
Values are coded as 0, 1, 2 representing the number of minor alleles.
}
\keyword{datasets}
