% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_phenotype}
\alias{example_phenotype}
\title{Example Phenotype Data}
\format{
A data frame with 20 rows and 4 variables:
\describe{
  \item{FID}{Family ID}
  \item{IID}{Individual ID}
  \item{time}{Survival time in days}
  \item{status}{Event status (0 = censored, 1 = event)}
}
}
\source{
Simulated data for package examples
}
\usage{
example_phenotype
}
\description{
Time-to-event phenotype data for the example samples.
}
\keyword{datasets}
