% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_positions}
\alias{example_positions}
\title{Example SNP Positions}
\format{
A numeric vector of length 50
}
\source{
Simulated data for package examples
}
\usage{
example_positions
}
\description{
A vector of genomic positions (in base pairs) for the example SNPs.
}
\keyword{datasets}
