% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_analysis.R
\name{fit_null_model}
\alias{fit_null_model}
\title{Fit Null Cox Model}
\usage{
fit_null_model(time, status, covariates, formula = NULL, use_spacox = FALSE)
}
\arguments{
\item{time}{Vector of survival times}

\item{status}{Vector of event indicators (0 = censored, 1 = event)}

\item{covariates}{Matrix or data frame of covariates}

\item{formula}{Optional formula for the model. If NULL, uses all covariates.}

\item{use_spacox}{Logical, whether to use SPACox for model fitting. 
Requires SPACox package to be installed. Default is FALSE.}
}
\value{
A fitted Cox model object
}
\description{
Fit a Cox proportional hazards model with only covariates (no genetic variants).
This function supports both standard coxph and SPACox for large-scale genetic data.
SPACox (https://github.com/WenjianBI/SPACox) is recommended for genome-wide 
association studies with survival outcomes.
}
\note{
For large-scale genetic studies, it is recommended to use SPACox 
  by setting use_spacox = TRUE. SPACox can be installed from: 
  https://github.com/WenjianBI/SPACox
}
\examples{
\dontrun{
# Fit null model with standard coxph
null_model <- fit_null_model(
  time = pheno_data$time,
  status = pheno_data$status,
  covariates = pheno_data[, c("age", "sex", "bmi")]
)

# Fit null model with SPACox (if available)
null_model_spa <- fit_null_model(
  time = pheno_data$time,
  status = pheno_data$status,
  covariates = pheno_data[, c("age", "sex", "bmi")],
  use_spacox = TRUE
)
}

}
