% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{Get.cauchy}
\alias{Get.cauchy}
\title{Cauchy Combination Function}
\usage{
Get.cauchy(p)
}
\arguments{
\item{p}{Vector of p-values}
}
\value{
Combined p-value using Cauchy combination
}
\description{
Cauchy Combination Function
}
\keyword{internal}
