% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{Get.cauchy.scan}
\alias{Get.cauchy.scan}
\title{Cauchy Combination for Sliding Windows}
\usage{
Get.cauchy.scan(p, window.matrix)
}
\arguments{
\item{p}{Vector of p-values}

\item{window.matrix}{Window indicator matrix}
}
\value{
Vector of combined p-values
}
\description{
Cauchy Combination for Sliding Windows
}
\keyword{internal}
