% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{Get.p}
\alias{Get.p}
\title{Get P-values using Score Test}
\usage{
Get.p(X, result.prelim)
}
\arguments{
\item{X}{Genotype matrix}

\item{result.prelim}{Preliminary results from KS.prelim}
}
\value{
Vector of p-values
}
\description{
Get P-values using Score Test
}
\keyword{internal}
