% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{Impute}
\alias{Impute}
\title{Imputation Function}
\usage{
Impute(Z, impute.method)
}
\arguments{
\item{Z}{Matrix with missing values}

\item{impute.method}{Imputation method ("fixed" or "mean")}
}
\value{
Imputed matrix
}
\description{
Imputation Function
}
\keyword{internal}
