% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filter.R
\name{knockoff-filter}
\alias{knockoff-filter}
\title{Knockoff Filter and Variable Selection}
\description{
Functions for computing W statistics and applying knockoff filters
for variable selection with false discovery rate control.
}
