% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_knockoff_gds}
\alias{load_knockoff_gds}
\title{Load Knockoff Data from GDS File}
\usage{
load_knockoff_gds(gds_file)
}
\arguments{
\item{gds_file}{Path to the GDS file containing knockoff data}
}
\value{
List containing:
  \item{original}{Original genotype matrix}
  \item{knockoffs}{List of knockoff matrices}
  \item{sample_ids}{Sample IDs}
  \item{positions}{SNP positions}
}
\description{
Load original genotypes and knockoff variables from a GDS file.
}
\keyword{internal}
