% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_plink_data}
\alias{load_plink_data}
\title{Load PLINK Format Genetic Data}
\usage{
load_plink_data(bed_path, verbose = TRUE)
}
\arguments{
\item{bed_path}{Path to PLINK .bed file (without extension)}

\item{verbose}{Whether to print loading progress (default: TRUE)}
}
\value{
List containing genotype matrix and metadata
}
\description{
Loads genetic data from PLINK binary format (.bed/.bim/.fam files) and
converts it to a sparse matrix format suitable for knockoff analysis.
}
\examples{
\dontrun{
# Load PLINK data
extdata_path <- system.file('extdata', package = 'CoxMK')
plink_data <- load_plink_data(file.path(extdata_path, 'sample'))
}
}
\keyword{internal}
