% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{max.nth}
\alias{max.nth}
\title{Model-X Knockoff Statistics and Fil#' @return The n-th largest value in the vector}
\usage{
\method{max}{nth}(x, n)
}
\description{
Model-X Knockoff Statistics and Fil#' @return The n-th largest value in the vector
}
