% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{max_nth}
\alias{max_nth}
\title{Find N-th Largest Value}
\usage{
max_nth(x, n)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{The position (1 = largest, 2 = second largest, etc.)}
}
\value{
The n-th largest value in the vector
}
\description{
Returns the n-th largest value in a vector
}
