% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{MK.q.byStat}
\alias{MK.q.byStat}
\title{Calculate Model-X Knockoff Q-values by Statistics}
\usage{
MK.q.byStat(kappa, tau, M, Rej.Bound = 10000)
}
\arguments{
\item{kappa}{Vector of kappa statistics}

\item{tau}{Vector of tau statistics}

\item{M}{Number of knockoff copies}

\item{Rej.Bound}{Maximum number of rejections}
}
\value{
Vector of q-values
}
\description{
Calculate Model-X Knockoff Q-values by Statistics
}
