% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{MK.statistic}
\alias{MK.statistic}
\title{Calculate Model-X Knockoff Statistics}
\usage{
MK.statistic(T_0, T_k, method = "median")
}
\arguments{
\item{T_0}{Test statistics for original variables}

\item{T_k}{Test statistics for knockoff variables (matrix)}

\item{method}{Method for calculating statistics ("median" or "max")}
}
\value{
Matrix with kappa and tau statistics
}
\description{
Calculate Model-X Knockoff Statistics
}
\keyword{internal}
