% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{MK.threshold.byStat}
\alias{MK.threshold.byStat}
\title{Calculate Model-X Knockoff Threshold by Statistics}
\usage{
MK.threshold.byStat(kappa, tau, M, fdr = 0.1, Rej.Bound = 10000)
}
\arguments{
\item{kappa}{Vector of kappa statistics}

\item{tau}{Vector of tau statistics}

\item{M}{Number of knockoff copies}

\item{fdr}{Target false discovery rate}

\item{Rej.Bound}{Maximum number of rejections}
}
\value{
Threshold value
}
\description{
Calculate Model-X Knockoff Threshold by Statistics
}
\keyword{internal}
