% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_statistics.R
\name{MK.threshold}
\alias{MK.threshold}
\title{Calculate Model-X Knockoff Threshold}
\usage{
MK.threshold(T_0, T_k, fdr = 0.1, method = "median", Rej.Bound = 10000)
}
\arguments{
\item{T_0}{Test statistics for original variables}

\item{T_k}{Test statistics for knockoff variables}

\item{fdr}{Target false discovery rate}

\item{method}{Method for calculating statistics}

\item{Rej.Bound}{Maximum number of rejections}
}
\value{
Threshold value
}
\description{
Calculate Model-X Knockoff Threshold
}
