% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filter.R
\name{mk_threshold_by_stat}
\alias{mk_threshold_by_stat}
\title{Calculate Model-X Knockoff Threshold by Statistics}
\usage{
mk_threshold_by_stat(kappa, tau, M, fdr = 0.1, Rej.Bound = 10000)
}
\arguments{
\item{kappa}{Vector of kappa statistics}

\item{tau}{Vector of tau statistics}

\item{M}{Number of knockoff copies}

\item{fdr}{Target false discovery rate}

\item{Rej.Bound}{Maximum number of rejections}
}
\value{
Threshold value
}
\description{
Calculate Model-X Knockoff Threshold by Statistics
}
\keyword{internal}
