% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{read_bed_simple}
\alias{read_bed_simple}
\title{Simple BED file reader (fallback when BEDMatrix is not available)}
\usage{
read_bed_simple(bed_file, n_samples, n_snps)
}
\arguments{
\item{bed_file}{Path to .bed file}

\item{n_samples}{Number of samples}

\item{n_snps}{Number of SNPs}
}
\value{
Matrix of genotypes
}
\description{
Simple BED file reader (fallback when BEDMatrix is not available)
}
\keyword{internal}
