% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knockoffs.R
\name{sparse_cor}
\alias{sparse_cor}
\title{Helper function for sparse correlation calculation}
\usage{
sparse_cor(x)
}
\arguments{
\item{x}{A sparse matrix}
}
\value{
A list containing covariance and correlation matrices
}
\description{
Helper function for sparse correlation calculation
}
\keyword{internal}
