% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knockoffs.R
\name{sparse_cov_cross}
\alias{sparse_cov_cross}
\title{Helper function for sparse cross-covariance calculation}
\usage{
sparse_cov_cross(x, y)
}
\arguments{
\item{x}{A sparse matrix}

\item{y}{A sparse matrix}
}
\value{
Cross-covariance matrix
}
\description{
Helper function for sparse cross-covariance calculation
}
\keyword{internal}
