% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_plink_data}
\alias{load_plink_data}
\title{Load PLINK Format Data}
\usage{
load_plink_data(bed_file, verbose = TRUE)
}
\arguments{
\item{bed_file}{Path to the .bed file (without extension, the function 
will look for .bed, .bim, and .fam files with the same prefix).}

\item{verbose}{Whether to print loading progress (default: TRUE).}
}
\value{
A list containing:
  \item{genotypes}{Sparse matrix of genotypes (samples x SNPs)}
  \item{fam}{Data frame with family information}
  \item{bim}{Data frame with SNP information}
  \item{positions}{Vector of SNP positions}
  \item{sample_ids}{Vector of sample IDs}
}
\description{
Load genotype data from PLINK binary format files (.bed/.bim/.fam).
}
\examples{
\dontrun{
# Load PLINK data
plink_data <- load_plink_data("path/to/data")
}

}
