% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{prepare_phenotype}
\alias{prepare_phenotype}
\title{Prepare Phenotype Data}
\usage{
prepare_phenotype(pheno_file, covar_file, sample_ids)
}
\arguments{
\item{pheno_file}{Path to phenotype file with columns: FID, IID, time, status}

\item{covar_file}{Path to covariate file with columns: FID, IID, age, sex, etc.}

\item{sample_ids}{Vector of sample IDs to match with genotype data}
}
\value{
A data frame with merged phenotype and covariate data
}
\description{
Load and prepare time-to-event phenotype data and covariates.
}
\examples{
\dontrun{
# Prepare phenotype data
pheno_data <- prepare_phenotype(
  "tte_phenotype.txt", 
  "covariates.txt", 
  sample_ids
)
}

}
